% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydough.R
\name{install_mido}
\alias{install_mido}
\title{Install \code{mido}}
\usage{
install_mido()
}
\value{
a \code{reticulate} \code{py_config} object
}
\description{
Creates the \code{mydough} conda environment and installs Python,
\code{numpy} and \code{mido} in it.
}
\details{
If \code{Miniconda} isn't installed, \code{install_mido} installs it in the
default location.
}
\examples{
\dontrun{
  config <- mydough::install_mido()
  print(config)
}

}
