% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydough.R
\name{new_midi_file}
\alias{new_midi_file}
\title{New MIDI file}
\usage{
new_midi_file(type = 1, ticks_per_beat = 480)
}
\arguments{
\item{type}{integer, MIDI file type, default is 1}

\item{ticks_per_beat}{integer, default is 480}
}
\value{
an empty MidiFile object
}
\description{
Create a new MidiFile object using \code{mido}
}
\details{
The "empty" MidiFile object from \code{mido} defaults to a type 1
MIDI file with \code{ticks_per_beat} set to 480. From
\insertCite{midoMidoMIDI}{mydough}, section 3 - "There are three types of
MIDI files:
\itemize{
\item type 0 (single track): all messages are saved in one track
\item type 1 (synchronous): all tracks start at the same time
\item type 2 (asynchronous): each track is independent of the others."
}
}
\examples{
\dontrun{
  print(midi_file_object <- mydough::new_midi_file())
}
}
\references{
\insertRef{midoMidoMIDI}{mydough}
}
