% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{install_mido}
\alias{install_mido}
\title{Install \code{mido}}
\usage{
install_mido()
}
\value{
the full path to the file
}
\description{
Creates a \code{mydough} conda environment and installs \code{mido} in it.
\code{Miniconda} will be installed if necessary.
}
