% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{read_midi_file}
\alias{read_midi_file}
\title{Read MIDI file}
\usage{
read_midi_file(midi_file)
}
\arguments{
\item{midi_file}{a character string with the path to a MIDI file. Note that
"~" does \emph{\strong{not}} work - you need to specify an absolute path!}
}
\value{
the MidiFile object read from the given file
}
\description{
Reads a MIDI file using \code{mido}
}
\examples{
\dontrun{
  reticulate::use_condaenv("mydough")
  midi_file_object <- mydough::read_midi_file(
    "/home/znmeb/Projects/eikosany/inst/Eikosany-MIDI-Files/eikosany-1-3-5-11.mid"
  )
  print(midi_file_object)
}

}
